1341. Movie Rating (09.04.2025)
I did this problem yesterday (09.04.2025), but I was busy as it was the day before 2 important exams. I forgot to upload it to GitHub and submit it, but I am doing it the next day so I don't think it caused much of an issue. The most important to me is education, not streak or a nice commit field. This problem used an extraction, it wasn't terrible but I did have to think about it for a little.